+++ src/ragger/embedder.py
@@ -51,18 +51,6 @@ class E5Embedder(Embedder):
         """The maximum length of the context that the embedder can handle."""
         return self.tokenizer.model_max_length
 
-    def tokenize(self, text: str | list[str]) -> np.ndarray:
-        """Tokenize a text.
-
-        Args:
-            text:
-                The text or texts to tokenize.
-
-        Returns:
-            The tokens of the text.
-        """
-        return self.tokenizer.tokenize(text)
-
     def embed_documents(self, documents: Iterable[Document]) -> list[Embedding]:
         """Embed a list of documents using an E5 model.