+++ tests/test_document_store.py
@@ -1,44 +1,59 @@
 """Unit tests for the `document_store` module."""
 
 import typing
-from pathlib import Path
-from tempfile import NamedTemporaryFile
 
 import pytest
-from ragger.document_store import DocumentStore, JsonlDocumentStore
+import ragger.document_store
+from ragger.data_models import DocumentStore
 
 
-class TestJsonlDocumentStore:
-    """Tests for the `JsonlDocumentStore` class."""
+@pytest.fixture(
+    scope="module",
+    params=[
+        cls
+        for cls in vars(ragger.document_store).values()
+        if isinstance(cls, type)
+        and issubclass(cls, DocumentStore)
+        and cls is not DocumentStore
+    ],
+)
+def document_store(
+    documents, request, special_kwargs
+) -> typing.Generator[DocumentStore, None, None]:
+    """Initialise a document store for testing."""
+    document_store_cls = request.param
+    document_store = document_store_cls(
+        **special_kwargs.get(document_store_cls.__name__, {})
+    )
+    document_store.add_documents(documents=documents)
+    yield document_store
+    document_store.remove()
 
-    @pytest.fixture(scope="class")
-    def document_store(
-        self, documents
-    ) -> typing.Generator[JsonlDocumentStore, None, None]:
-        """Initialise a JsonlDocumentStore for testing."""
-        with NamedTemporaryFile(mode="w", suffix=".jsonl") as file:
-            data_str = "\n".join(document.model_dump_json() for document in documents)
-            file.write(data_str)
-            file.flush()
-            yield JsonlDocumentStore(path=Path(file.name))
 
-    def test_is_document_store(self):
-        """Test that the JsonlDocumentStore is a DocumentStore."""
-        assert issubclass(JsonlDocumentStore, DocumentStore)
+def test_initialisation(document_store):
+    """Test that the document store can be initialised."""
+    assert isinstance(document_store, DocumentStore)
 
-    def test_initialisation(self, document_store):
-        """Test that the JsonlDocumentStore can be initialised."""
-        assert document_store
 
-    def test_getitem(self, document_store):
-        """Test that documents can be fetched from the JsonlDocumentStore."""
-        assert document_store["1"].text == "Den hvide kat hedder Sjusk."
-        assert document_store["2"].text == "Den sorte kat hedder Sutsko."
-        assert document_store["3"].text == "Den røde kat hedder Pjuskebusk."
-        assert document_store["4"].text == "Den grønne kat hedder Sjask."
-        assert document_store["5"].text == "Den blå kat hedder Sky."
+def test_getitem(document_store, documents):
+    """Test that documents can be fetched from the document store."""
+    for document in documents:
+        assert document_store[document.id] == document
 
-    def test_getitem_missing(self, document_store):
-        """Test that fetching a missing document raises a KeyError."""
-        with pytest.raises(KeyError):
-            document_store["6"]
+
+def test_getitem_missing(document_store, non_existing_id):
+    """Test that fetching a missing document raises a KeyError."""
+    with pytest.raises(KeyError):
+        document_store[non_existing_id]
+
+
+def test_contains(documents, document_store, non_existing_id):
+    """Test that the document store can check if it contains a document."""
+    for document in documents:
+        assert document.id in document_store
+    assert non_existing_id not in document_store
+
+
+def test_len(document_store, documents):
+    """Test that the document store can return the number of documents."""
+    assert len(document_store) == len(documents)