+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+- Added a `SqliteDocumentStore` that uses a SQLite database to store documents.
+
+### Changed
+- Added defaults to all arguments in each component's constructor, so that the
+  user can create a component without specifying any arguments. This also allows for
+  uniform testing of all components.
+
+
 ## [v1.0.0] - 2024-08-12
 ### Added
 - Initial release, with the document store `JsonlDocumentStore`, the embedder